+++ src/sys/mips/mod.rs
@@ -4,7 +4,7 @@
 //!
 //! Refs:
 //! - Unified Hosting Interface. MD01069 Reference Manual <https://docplayer.net/62444141-Unified-hosting-interface-md01069-reference-manual.html>
-//! - <https://github.com/qemu/qemu/blob/HEAD/target/mips/tcg/sysemu/mips-semi.c>
+//! - <https://github.com/qemu/qemu/blob/v9.1.0/target/mips/tcg/sysemu/mips-semi.c>
 
 #![allow(clippy::missing_safety_doc)] // TODO